+++ internal/models/user_chest_history.go
@@ -7,7 +7,7 @@ import (
 type UserChestHistory struct {
 	ID          uint      `gorm:"primaryKey" json:"id"`                     // Уникальный идентификатор
 	UserChestID uint      `gorm:"not null" json:"user_id"`                  // Ссылка на пользователя
-	Health      uint      `gorm:"not null" json:"health"`                   // Текущее здоровье сундука
+	Health      uint64    `gorm:"not null" json:"health"`                   // Текущее здоровье сундука
 	Level       int       `gorm:"default:1" json:"level"`                   // Уровень сундука
 	Amount      uint32    `json:"amount"`                                   // Уровень сундука
 	CreatedAt   time.Time `gorm:"autoCreateTime" json:"created_at"`         // Время создания
